requirements.txt
APScheduler==3.6.3
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.7
django-apscheduler==0.3.0
django-ranged-response==0.2.0
django-ratelimit==2.0.0
django-simple-captcha==0.5.12
gunicorn==20.0.4
idna==2.8
maxminddb==1.5.2
Pillow==6.2.1
psycopg2==2.8.4
pycodestyle==2.5.0
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
tzlocal==2.0.0
urllib3==1.25.7
whitenoise==5.0.1
setuptools==41.2.0